fixup: build fixes for android-tv-14.0.0_r1 base
Three pre-existing breakages in the android-tv-14.0.0_r1 import prevent
any board from linking. They are not caused by the rebase - a pristine
android-tv-14.0.0_r1 checkout of an untouched upstream board
(sm1_ac214_v1) fails with the identical errors.

1. common/cmd_avb.c: avb2_kpub_{vendor,default}_len are declared
   'const size_t' but defined 'const int' in every board avb2_kpub.c
   and in lib/libavb/testkey.c. The 64-bit load emitted for the size_t
   declaration needs 8-byte alignment, so the 4-byte-aligned int symbol
   gives "relocation truncated to fit: R_AARCH64_LDST64_ABS_LO12_NC".
   Match the declarations to the definitions.

2. common/Makefile: upstream commented out the unconditional
   'obj-y += env_amlnand.o' in the CONFIG_STORE_COMPATIBLE branch, but
   common/env_storage.c still calls amlnand_saveenv() and
   amlnand_env_relocate_spec() when CONFIG_AML_MTD/CONFIG_AML_NAND is
   set, giving undefined references. Restore the rule.

3. board/amlogic/configs/{g12a_bananapim5,g12a_odroidc4,g12a_radxa0,
   sm1_bananapim5}_v1.h: include/amlogic/aml_cec.h now defaults
   CONFIG_CEC_OSD_NAME to CONFIG_PRODUCT_NAME instead of "AML_TV". In
   the bl301/scp_task build aml_cec.h is included before the board
   config, so the board's "AML_TV" is now a differing redefinition and
   -Werror rejects it. #undef before defining.

With these, sm1_bananapim5_v1, g12a_bananapim5_v1, g12a_odroidc4_v1 and
g12a_radxa0_v1 all build to completion (u-boot.bin, bl301.bin, acs.bin)
using the Linaro 4.8-2013.11 aarch64-none-elf + arm-none-eabi
toolchains, with CROSS_COMPILE_T32=arm-none-eabi-.

+++ common/cmd_avb.c
@@ -64,13 +64,13 @@ extern const size_t avb2_kpub_vendor_external_len[];
 extern const size_t avb2_kpub_vendor_external_num;
 #else
 extern const uint8_t avb2_kpub_vendor[];
-extern const size_t avb2_kpub_vendor_len;
+extern const int avb2_kpub_vendor_len;
 #endif /* CONFIG_AVB2_KPUB_VENDOR_MULTIPLE */
 #endif /* CONFIG_AVB_KPUB_VENDOR */
 
 #if defined(CONFIG_AVB2_KPUB_DEFAULT) || defined(CONFIG_AVB2_KPUB_DEFAULT_VENDOR)
 extern const uint8_t avb2_kpub_default[];
-extern const size_t avb2_kpub_default_len;
+extern const int avb2_kpub_default_len;
 #endif /* CONFIG_AVB_KPUB_DEFAULT || CONFIG_AVB2_KPUB_DEFAULT_VENDOR */
 
 AvbOps avb_ops_;
